Add completion handling
